Update symbol for Indian Rupee
Source : https://en.wikipedia.org/wiki/Indian_rupee_sign

The Indian rupee sign (sign: ₹; code: INR) is the currency sign for the Indian rupee, the official currency of India. Designed by Udaya Kumar, it was presented to the public by the Government of India on 15 July 2010,[1] following its selection through an “open” competition among Indian residents. Before its adoption, the most commonly used symbols for the rupee were Rs, Re or, if the text was in an Indian language, an appropriate abbreviation in that language.

The design is based on the Devanagari letter "र" (ra) with a double horizontal line at the top. It also resembles the Latin capital letter "R", especially R rotunda (Ꝛ).

The Unicode character for the Indian rupee sign is U+20B9 ₹ Indian Rupee Sign.